ib_core: removed netlink support
netlink support was implemented by NirM and the code is in upstream.

It's a basic RDMA netlink infrastructure that allows for registration of RDMA clients for which data is to be exported and supplies message
construction callbacks.

Consulted with MoniS (rdma_cm) who said that as for this day there's no use for it in user space.

FreeBSD has no netlink support.

issue: 378274

Change-Id: If012c495f2de0e568a61031abdfd1d5ffdf0100b
Signed-off-by: Meny Yossefi <menyy@mellanox.com>