en_tx.c: removed data frags backward logic
In Linux, we start DMA mapping from the last data segment and backwards.
This is because the HW pre-fetches the data when sending a packet.
If the first data frag is 0xffff, the HW knows not the fetch this data frag.
No way to do so in FreeBSD since unlike in Linux where the data frags are in an array,
in FreeBSD data frag mbufs are in a one sided linked list.

Change-Id: I254df5b199f54cdb5b8493f9545c671fd844c207
Signed-off-by: Meny Yossefi <menyy@mellanox.com>