mlx4_en: Prevention of CQ overrun while transmiting traffic
In the packet xmit process as it is today, mlx4_en_xmit is called
for each outgoing packet.
At the end of the transmit procedure just before writing the packet
to the HW, the driver checks whether the qp has enough free space to hold the
relevant wqe. If the qp has space for only one more wqe (the wqe in question),
it blocks itself and the OS from getting anymore sending requests.
The problem occurs when running in multithreaded enviourment
and anothe thread is already waiting on the qp's lock,
in order to send another packet through mlx4_en_xmit.
In this case the thread will go in the function and send another outgoing packet,
although the qp is full.
This will result with a cq overrun.

Our change moves the space check to the beginning of the xmit function,
exit the function in case the qp is full and marking the queue as blocked.
This way all waiting threads will exit the function as well and will
not cause memory leakage.

Two importent facts to be aware of:
First, the fuction calling mlx4_en_xmit checks the return value, and in case of
an error (like exiting before completing the transmit) it will recall the function
for the same packet - what will prevent us from losing unsent packets.
Second, before exiting the function, the driver arming the cq.

issue: 402622

Change-Id: I7fd370565033f69c5d9c9eab0a5e96300c375d74
Signed-off-by: Shani Michaeli <shanim@mellanox.com>