mlxen fix: crash when adding interface in down state to lagg
consult priv->port_up before notifying the OS that the port is up.
on init_netdev we call INIT_PORT which in turn fire an PORT_UP event.
linkstate() is handling the event and should not notify the OS if the
port is not ready to traffic(via start_port).
Since we might not get the event again on port up. we will notify
the OS in set_rx_mode which is called during start_port.

Issue: 409048

Change-Id: Ic050d1c2c2e222918acbd599410f48ca63619f45
Signed-off-by: Shahar Klein <shahark@mellanox.com>