mlxen: unloading mlxen module while sending traffic in background
When unloading the driver, we unregister the device before
calling the stop port function and release our working buffers
including the tx buff.

When unloading while traffic is being sent by the sender,
all the packets which were already collected in the send queue are
being transmited to the receiver at once.

Currently, while freeing the tx descriptor, each packet is copied
to the BPF listener before its mbuf is released. This process
requires the use of the device and its fields.

This sequence of events causes the driver to crash while using
a device which was already unregistered.
In order to prevent it, we make sure the device exsists before copying the
packet, and if not, we will not go through with it.

issue: 409210

Change-Id: I1e4f1e88b6a37669216ae6fb3b8d19d6318f9884
Signed-off-by: Shani Michaeli <shanim@mellanox.com>