mlx4ib Integration: Removed pdev->bus reference.
The call to sprintf with the argument dev->pdev->bus->conf.pd_name
caused a crash while changing a port type from ETH to IB.
In Linux this argument is the pci bus which is a part of the eq name in
the function mlx4_ib_alloc_eqs, In FreeBSD the bus struct is NULL, this
struct has never been initialized. Removed this struct from the pci_dev
struct in pci.h file, and Replaced this argument with four FBSD kernel getters:
	- pci_get_domain
	- pci_get_bus
	- pci_get_dev
	- pci_get_func
which together create the pci.
In addition, removed redundant bus struct usage in mlx4_core main.c

Issue : 417960

Change-Id: Ia01410b3d45132cd2a2a24454dc4fe8b0ff7c8bd
Signed-off-by: Daria Genzel <dariaz@mellanox.com>