ibcore/mlx4ib : Fix - ibtools: broken ibstat
ibstat was broken because the returned value from the function
show_port_gid was corrupted.
The returned gid value was built using the sprintf func which used the %pI6 format
which is working in Linux but not working in FreeBSD.
Replaced this format and it's arguments with macros of the working format (can be founf in 10-RELEASE) and
the sutable args.
Replaced all %pI6 instances in all modules except of ipoib.

Issue : 422614
Issue : 431346

Change-Id: Ieaa272de1df1ce190ab2c2d7f4280b00848f996d
Signed-off-by: Daria Genzel <dariaz@mellanox.com>