Make pr_* macros use freebsd log system.
FreeBSD doesn't use printk, this commit changes pr_* macros
to use the log system in freebsd.

this commit changes only the behaviour of pr_* macros.
legacy code that still calls printk directly should be changed in the future.
in order not to effact that code, we keep the macro compatibility layer of printk -> printf

Change-Id: I0890250b232cf033e54d30e4990e8c67ea013da9
Signed-off-by: Mark Bloch <markb@mellanox.com>