ipoib: Fix - 'multicast join failed' message appears in the dmesg
ipoib starts the mcast_join task whenever the interface is "up" without
connection to the ib port state, and many requests sent to the sm but
since the port state is not ACTIVE for all of them it gets error (-22).
In order to avoid it I added a check to the port state in the
mcast_join_task, and only if the port is in ACTIVE state run the task,
if not the function returns, till new event will come
(IB_EVENT_PORT_ACTIVE or IB_EVENT_CLIENT_REREGISTER) and the task will
run again.

Issue: 439735

Change-Id: I69430d0325deb13524a4a5ba28514a105c4ca0e0
Signed-off-by: Daria Genzel <dariaz@mellanox.com>