ipoib: Fix - crash while unloading the ipoib module
The next scenario causes a crash: Load ipoib, configure at least one ip
and unload the ipoib module.

This crash happens because some of the mbufs were released twice, Once from the ipoib_poll
and the second time from the ipoib_drain_cq.
In ipoib_ib.c usually the function ipoib_poll called as a result of an interrupt
invoked by getting a complition. In our case when uloading the ipoib module
the same interrupt invokes the ipoib_poll because while unloading the module
we change the state of the QPs to an error state what sends a complition also.
In the ipoib_poll there is an ib_poll_cq call and ipoib_ib_handle_rx_wc call which
deletes the mbufs. In the same time the ipoib_drain_cq called from the ipoib_ib_dev_stop
which invoked because we unloaded the module. In the ipoib_drain_cq there is an
ib_poll_cq call and ipoib_ib_handle_rx_wc call also.
These two m_freem without locking caused the crash.
In Linux this crash won't happen because in drain_cq they take the local_bh_disable lock
which is used by the NAPI (triggers the poll) what makes it a safe locking.

Change-Id: Ib8a072d6c066475af36370cb86ecb571fd0b0264
Signed-off-by: Daria Genzel <dariaz@mellanox.com>