New WQE format.
There is a known (and old) bug in the firmware when reading WQEs.
A workaround was implemented long ago to overcome said bug, although the fix worked
there were cases that when TSO was used, we saw performance degradation.

The new WQE changes the way driver-firmware talk to each other which in turn fixes the TSO issue.

This patch is based on the work of: Haggai Abramovsky

The original patchs:

http://l-gerrit.mtl.labs.mlnx:8080/55817
http://l-gerrit.mtl.labs.mlnx:8080/55994

Change-Id: Ia7ec2778d178bc3f9f5d29b47845f2c252a51900
Signed-off-by: Mark Bloch <markb@mellanox.com>