Packet Pacing: RL ring gets full and stops traffic on native rings
When rl ring buf is full it stops both the ring and the queue (port).
In Packet Pacing rings it is expected that rings will get full so
we don't want them to stop the port.
Additionally - the drbr size should be smaller than the QP. The rate
limited qp will get full and there is no need for a big drbr to hold
the packets sent by the above stack, increasing memory stamp for no
reason.
The fix:
1. In xmit() - block port only if this is native ring
2. In transmit_locked() - don't xmit also if ring->blocked == 1
3. In create_ring() - rl_ring drbr size = ring_size / 2

Issue: 550341

Change-Id: Ifca67ee69d91887b59ea4021bde9e675163c42c3
Signed-off-by: Oded Shanoon <odeds@mellanox.com>