Packet Pacing - Fix a memory leak.
There is a memory leak because when we run mlx4_en_stop_port we check if the ring is valid
(not under the ring lock), but in destroy/create/modify we first change status of ring
and only then check if port is up.

There is an inconsistency in the checks and a race condition can accure.

Issue : 553390

Change-Id: I19719be2cef71436e99c221b382c5b5eb00942d4
Signed-off-by: Mark Bloch <markb@mellanox.com>