Updated freebsd/ofed_freebsd_3_0 Project: drivex  fa3864d0f71a1383170992b74120c6e5d0833135
[Windows Only] Post send return with immediate error when the corresponding QP is in RESET/INIT/RTR state.
This is done according to the following sections from IB spec:
C10-22: If a Work Request is submitted to a Work Queue while its corresponding
QP is in the Reset State, an immediate error shall be returned.

C10-25: If a Work Request is submitted to a Send Queue while its corresponding
QP is in the Init State, an immediate error shall be returned.

C10-29: If a Work Request is submitted to a Send Queue while its corresponding
QP is in the RTR State, an immediate error shall be returned.

Change-Id: If073b268688d85913edb7967311d32bbc5fd7fad
Signed-off-by: Tamer Hleihel <tamerh@mellanox.com>